pyproject.toml
[project]
name = "mcu-purple-baseline"
version = "0.1.0"
description = "A2A-compatible purple baseline agent for MCU AgentBeats benchmark"
requires-python = ">=3.10"

dependencies = [
  "a2a-sdk>=0.3.0",
  "uvicorn>=0.23",
  "starlette>=0.37",
  "httpx>=0.25",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "numpy>=1.24",
  "opencv-python-headless>=4.8",
  "pytest>=7.0",
  "pytest-asyncio>=0.21",
]

[project.scripts]
mcu-purple = "server.app:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
